+++ codec/CHANGES.md
@@ -4,6 +4,8 @@
 
 * Remove `Packet::Empty`
 
+* Add `max frame size` config
+
 ## [0.1.0] - 2019-06-18
 
 * Initial release